+++ mods/berlords-better-horses/src/main/java/io/github/bertie_mc/betterhorses/HorseEquipment.java
@@ -8,9 +8,15 @@ public interface HorseEquipment {
 
     SimpleContainer betterhorses$saddleInventory();
 
+    SimpleContainer betterhorses$storage();
+
     ItemStack betterhorses$syncedShoes();
 
     ItemStack betterhorses$syncedSaddle();
+
+    default boolean betterhorses$traveller() {
+        return betterhorses$syncedSaddle().is(BetterHorses.WANDERER.get());
+    }
 
     default ShoeTier betterhorses$tier() {
         return betterhorses$syncedShoes().getItem() instanceof HorseshoeItem item ? item.tier : ShoeTier.NONE;